Fill out test view.
testViewCtrl.js:
- Added a field to the test data that holds the raw datetiem data, not
the existing formatted timeStamp string field. 

testViewDirective.js / testViewTemplate.js:
- Enhanced to show Test Details, History and mocked out Compare
function. The selection works to populate the compare but clicking teh
Compare linl will not navigate anywhere exciting.

Modify cal-heatmap to supress hovers. Will enhance this to provide rich
hover on test. Should do it in such a way as to be able to submit a pull
request.