Fix a couple of presentation bugs:
latestResultsGridDirective.js:
- getLatestResults function needs to break out of loop when the # latest
Results is less than the limit (5). Currently breaks as we'll get an
undefined related error as the in the case where for example there are
only 2 recent results. The previous code would still attemp to iterate
over the 3rd result.

testViewDirective.js:
- date should be initialised from the timestamp field of the history
item, not the dt field.

TODO: dt field  can probably be removed. Investigate where it is used.